update: add Repository class and move fetch functions from fragment
