+++ app/src/main/java/com/example/project1/ui/phonebook/PhoneBookFragment.java
@@ -57,93 +57,20 @@ public class PhoneBookFragment extends Fragment {
         switch (requestCode) {
             case PERMISSIONS_REQUEST_READ_CONTACTS:
                 if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED)
-                    getContactList();
+                    phoneBookViewModel.initializeContacts();
         }
     }
 
     private void requestContactList() {
-        if (ActivityCompat.checkSelfPermission(getContext(), android.Manifest.permission.READ_CONTACTS) == PackageManager.PERMISSION_GRANTED)
-            getContactList();
+        if (ActivityCompat.checkSelfPermission(getContext(), android.Manifest.permission.READ_CONTACTS) == PackageManager.PERMISSION_GRANTED) {
+            ArrayList<JsonData> data = phoneBookViewModel.getContacts().getValue();
+            if (data == null)
+                phoneBookViewModel.initializeContacts();
+            else
+                adapter.updateItems(phoneBookViewModel.getContacts().getValue());
+        }
         else
-            requestPermissions(new String[]{  Manifest.permission.READ_CONTACTS }, PERMISSIONS_REQUEST_READ_CONTACTS);
-    }
-
-    private String fetchPhoneNumber(ContentResolver cr, String id) {
-        Cursor phoneCursor = cr.query(
-                ContactsContract.CommonDataKinds.Phone.CONTENT_URI,
-                null,
-                ContactsContract.CommonDataKinds.Phone.CONTACT_ID + " = " + id,
-                null,
-                null
-        );
-        String number = "";
-
-        if (phoneCursor.moveToFirst())
-            number = phoneCursor.getString(phoneCursor.getColumnIndex(ContactsContract.CommonDataKinds.Phone.NUMBER));
-
-        phoneCursor.close();
-        return number;
-    }
-
-    private String fetchEmail(ContentResolver cr, String id) {
-        Cursor emailCursor = cr.query(
-                ContactsContract.CommonDataKinds.Email.CONTENT_URI,
-                null,
-                ContactsContract.CommonDataKinds.Email.CONTACT_ID + " = " + id,
-                null,
-                null
-        );
-
-        String email = "";
-        if (emailCursor.moveToFirst())
-            email = emailCursor.getString(emailCursor.getColumnIndex(ContactsContract.CommonDataKinds.Email.DATA));
-
-        emailCursor.close();
-        return email;
-    }
-
-    private Uri fetchPhotoUri(ContentResolver cr, String id) {
-        try {
-            Cursor cursor = cr.query(
-                    ContactsContract.Data.CONTENT_URI,
-                    null,
-                    ContactsContract.Data.CONTACT_ID + " = " + id + " AND " + ContactsContract.Data.MIMETYPE + "='" + ContactsContract.CommonDataKinds.Phone.CONTENT_ITEM_TYPE + "'",
-                    null,
-                    null
-            );
-            if (cursor == null || !cursor.moveToFirst())
-                return null;
-
-        } catch (Exception e) {
-            e.printStackTrace();
-            return null;
-        }
-
-        Uri person = ContentUris.withAppendedId(ContactsContract.Contacts.CONTENT_URI, Long.parseLong(id));
-        return Uri.withAppendedPath(person, ContactsContract.Contacts.Photo.CONTENT_DIRECTORY);
-    }
-
-    private void getContactList() {
-        ContentResolver cr = getActivity().getContentResolver();
-        Cursor cur = cr.query(ContactsContract.Contacts.CONTENT_URI, null, null, null, null);
-
-        if (cur == null || cur.getCount() == 0)
-            return;
-
-        while (cur != null && cur.moveToNext()) {
-            String id = cur.getString(cur.getColumnIndex(ContactsContract.Contacts._ID));
-            String name = cur.getString(cur.getColumnIndex(ContactsContract.Contacts.DISPLAY_NAME));
-            String number = fetchPhoneNumber(cr, id);
-            String email = fetchEmail(cr, id);
-            Uri photo = fetchPhotoUri(cr, id);
-
-            contactList.add(new JsonData(name, number, email, photo));
-        }
-
-        if (cur != null)
-            cur.close();
-
-        adapter.notifyDataSetChanged();
+            requestPermissions(new String[]{ Manifest.permission.READ_CONTACTS }, PERMISSIONS_REQUEST_READ_CONTACTS);
     }
 }
